+++ src/internal/ui/metadata/metadata.go
@@ -4,11 +4,8 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
-	"os/user"
 	"path/filepath"
 	"sort"
-	"strconv"
-	"syscall"
 
 	"github.com/barasher/go-exiftool"
 
@@ -81,22 +78,6 @@ func GetMetadata(filePath string, metadataFocussed bool, et *exiftool.Exiftool) 
 	meta := getMetaDataUnsorted(filePath, metadataFocussed, et)
 	sortMetadata(meta.data)
 	return meta
-}
-
-func getOwnerAndGroup(fileInfo os.FileInfo) (string, string) {
-	usr := ""
-	grp := ""
-	if stat, ok := fileInfo.Sys().(*syscall.Stat_t); ok {
-		uid := strconv.FormatUint(uint64(stat.Uid), 10)
-		gid := strconv.FormatUint(uint64(stat.Gid), 10)
-		if userData, err := user.LookupId(uid); err == nil {
-			usr = userData.Username
-		}
-		if groupData, err := user.LookupGroupId(gid); err == nil {
-			grp = groupData.Name
-		}
-	}
-	return usr, grp
 }
 
 func getMetaDataUnsorted(filePath string, metadataFocussed bool, et *exiftool.Exiftool) Metadata {
