+++ src/internal/ui/metadata/metadata_windows.go
@@ -1,0 +1,11 @@
+//go:build windows
+
+package metadata
+
+import (
+	"os"
+)
+
+func getOwnerAndGroup(_ os.FileInfo) (string, string) {
+	return "", ""
+}
